+++ docs/documentation/build.md
@@ -366,3 +366,13 @@ Note: service worker support is experimental and subject to change.
     Use file name for lazy loaded chunks.
   </p>
 </details>
+
+<details>
+  <summary>bundle-dependencies</summary>
+  <p>
+    <code>--bundle-dependencies</code>
+  </p>
+  <p>
+    In a server build, state whether `all` or `none` dependencies should be bundles in the output.
+  </p>
+</details>
