+++ packages/@angular/cli/models/webpack-configs/browser.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs';
 import * as webpack from 'webpack';
 import * as path from 'path';
+import * as ts from 'typescript';
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const SubresourceIntegrityPlugin = require('webpack-subresource-integrity');
 
@@ -67,7 +68,16 @@ export function getBrowserConfig(wco: WebpackConfigOptions) {
     }));
   }
 
+  const supportES2015 = wco.tsConfig.options.target !== ts.ScriptTarget.ES3
+                     && wco.tsConfig.options.target !== ts.ScriptTarget.ES5;
+
   return {
+    resolve: {
+      mainFields: [
+        ...(supportES2015 ? ['es2015'] : []),
+        'browser', 'module', 'main'
+      ]
+    },
     output: {
       crossOriginLoading: buildOptions.subresourceIntegrity ? 'anonymous' : false
     },
@@ -91,6 +101,19 @@ export function getBrowserConfig(wco: WebpackConfigOptions) {
         minChunks: Infinity,
         name: 'inline'
       })
-    ].concat(extraPlugins)
+    ].concat(extraPlugins),
+    node: {
+      fs: 'empty',
+      // `global` should be kept true, removing it resulted in a
+      // massive size increase with Build Optimizer on AIO.
+      global: true,
+      crypto: 'empty',
+      tls: 'empty',
+      net: 'empty',
+      process: true,
+      module: false,
+      clearImmediate: false,
+      setImmediate: false
+    }
   };
 }
