+++ packages/@angular/cli/models/webpack-configs/server.ts
@@ -1,18 +1,32 @@
 import { WebpackConfigOptions } from '../webpack-config';
+import * as ts from 'typescript';
 
 /**
  * Returns a partial specific to creating a bundle for node
- * @param _wco Options which are include the build options and app config
+ * @param wco Options which are include the build options and app config
  */
-export function getServerConfig(_wco: WebpackConfigOptions) {
-  return {
+export function getServerConfig(wco: WebpackConfigOptions) {
+  const supportES2015 = wco.tsConfig.options.target !== ts.ScriptTarget.ES3
+                     && wco.tsConfig.options.target !== ts.ScriptTarget.ES5;
+
+  const config: any = {
+    resolve: {
+      mainFields: [
+        ...(supportES2015 ? ['es2015'] : []),
+        'main', 'module',
+      ],
+    },
     target: 'node',
     output: {
       libraryTarget: 'commonjs'
     },
-    externals: [
+    node: false,
+  };
+
+  if (wco.buildOptions.bundleDependencies == 'none') {
+    config.externals = [
       /^@angular/,
-      function (_: any, request: any, callback: (error?: any, result?: any) => void) {
+      (_: any, request: any, callback: (error?: any, result?: any) => void) => {
         // Absolute & Relative paths are not externals
         if (request.match(/^\.{0,2}\//)) {
           return callback();
@@ -33,6 +47,8 @@ export function getServerConfig(_wco: WebpackConfigOptions) {
           callback();
         }
       }
-    ]
-  };
+    ];
+  }
+
+  return config;
 }
